pyproject.toml
[tool.poetry]
name = "dougu"
version = "0.1.0"
description = ""
authors = ["ryuto konno <pxx.rai@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.7"
pytest = "6.2.5"
black = "21.12b0"
isort = "5.10.1"
mypy = "0.931"
logzero = "1.7.0"
tqdm = "4.62.3"
PyYAML = "6.0"
sklearn = "0.0"
mecab-python3 = "1.0.4"
ipadic = "1.0.0"
jaconv = "0.3"
sentencepiece = "0.1.96"
pandas = "1.4.0"
pandas-gbq = "0.17.0"
boto3 = "1.20.41"
google = "3.0.0"
google-auth = "2.4.0"
google-cloud-storage = "2.1.0"
seaborn = "0.11.2"
matplotlib = "3.5.1"
ipdb = "0.13.9"
ipython = "8.0.1"
types-PyYAML = "6.0.3"
python-dateutil = "^2.8.2"
types-python-dateutil = "^2.8.8"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
